untilize on device for bf16
