+++ runtime/include/tt/runtime/detail/ttnn/types.h
@@ -84,53 +84,6 @@ struct LayoutDesc {
   bool operator==(const LayoutDesc &other) const;
 };
 
-class LayoutConverter {
-public:
-  LayoutDesc inputDesc;
-  LayoutDesc outputDesc;
-  bool shouldTilize = false;
-  bool shouldUntilize = false;
-  bool shouldTypecast = false;
-  bool shouldToDevice = false;
-  bool shouldToMemoryConfig = false;
-  bool shouldFromDevice = false;
-
-  LayoutConverter(const LayoutDesc &inputDesc, const LayoutDesc &outputDesc);
-  ::ttnn::Tensor convertTensorLayout(const ::ttnn::Tensor &input,
-                                     OptionalMeshDeviceRef targetDevice);
-
-private:
-  ::ttnn::Tensor toLayoutIfNeeded(const ::ttnn::Tensor &input);
-  ::ttnn::Tensor typecastIfNeeded(const ::ttnn::Tensor &input);
-  ::ttnn::Tensor toDeviceIfNeeded(const ::ttnn::Tensor &input,
-                                  OptionalMeshDeviceRef targetDevice,
-                                  bool force = false);
-  ::ttnn::Tensor toMemoryConfigIfNeeded(const ::ttnn::Tensor &input);
-  ::ttnn::Tensor fromDeviceIfNeeded(const ::ttnn::Tensor &input);
-
-  ::ttnn::Tensor
-  handleHostInputNoLayoutNoTypecast(const ::ttnn::Tensor &input,
-                                    OptionalMeshDeviceRef targetDevice);
-  ::ttnn::Tensor
-  handleHostInputLayoutNoTypecast(const ::ttnn::Tensor &input,
-                                  OptionalMeshDeviceRef targetDevice);
-  ::ttnn::Tensor
-  handleHostInputNoLayoutTypecast(const ::ttnn::Tensor &input,
-                                  OptionalMeshDeviceRef targetDevice);
-  ::ttnn::Tensor
-  handleHostInputLayoutTypecast(const ::ttnn::Tensor &input,
-                                OptionalMeshDeviceRef targetDevice);
-  ::ttnn::Tensor convertHostTensorLayout(const ::ttnn::Tensor &input,
-                                         OptionalMeshDeviceRef targetDevice);
-
-  ::ttnn::Tensor
-  handleDeviceInputNoLayoutNoTypecast(const ::ttnn::Tensor &input);
-  ::ttnn::Tensor handleDeviceInputLayoutNoTypecast(const ::ttnn::Tensor &input);
-  ::ttnn::Tensor handleDeviceInputNoLayoutTypecast(const ::ttnn::Tensor &input);
-  ::ttnn::Tensor handleDeviceInputLayoutTypecast(const ::ttnn::Tensor &input);
-  ::ttnn::Tensor convertDeviceTensorLayout(const ::ttnn::Tensor &input);
-};
-
 class ProgramTensorPool {
 public:
   ProgramTensorPool(const std::vector<uint32_t> &programInputIds,
